db.validate_field_definitions
