+++ crates/apollo-compiler/src/validation/interface.rs
@@ -1,11 +1,11 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 
 use crate::{
     diagnostics::{
-        MissingField, OutputType, RecursiveDefinition, TransitiveImplementedInterfaces,
-        UndefinedDefinition, UniqueDefinition, UniqueField,
+        MissingField, RecursiveDefinition, TransitiveImplementedInterfaces, UndefinedDefinition,
+        UniqueDefinition,
     },
-    hir::{self, FieldDefinition},
+    hir,
     validation::{ast_type_definitions, ValidationSet},
     ApolloDiagnostic, ValidationDatabase,
 };
@@ -30,10 +30,6 @@ pub fn validate_interface_definition(
     interface_def: hir::InterfaceTypeDefinition,
 ) -> Vec<ApolloDiagnostic> {
     let mut diagnostics = Vec::new();
-
-    for field in interface_def.fields_definition() {
-        diagnostics.extend(db.validate_field_definition(field.clone()));
-    }
 
     // Interface definitions must have unique names.
     //
@@ -95,61 +91,7 @@ pub fn validate_interface_definition(
     }
 
     // Interface Type field validations.
-    let mut seen: HashMap<&str, &FieldDefinition> = HashMap::new();
-
-    let fields = interface_def.fields_definition();
-
-    for field in fields {
-        // Fields in an Interface definition must be unique
-        //
-        // Returns Unique Value error.
-        let field_name = field.name();
-        let offset = field.loc().offset();
-        let len = field.loc().node_len();
-
-        if let Some(prev_field) = seen.get(&field_name) {
-            let prev_offset = prev_field.loc().offset();
-            let prev_node_len = prev_field.loc().node_len();
-
-            diagnostics.push(ApolloDiagnostic::UniqueField(UniqueField {
-                field: field_name.into(),
-                src: db.source_code(prev_field.loc().file_id()),
-                original_field: (prev_offset, prev_node_len).into(),
-                redefined_field: (offset, len).into(),
-                help: Some(format!(
-                    "`{field_name}` field must only be defined once in this interface definition."
-                )),
-            }));
-        } else {
-            seen.insert(field_name, field);
-        }
-
-        // Field types in interface types must be of output type
-        if let Some(field_ty) = field.ty().type_def(db.upcast()) {
-            if !field.ty().is_output_type(db.upcast()) {
-                diagnostics.push(ApolloDiagnostic::OutputType(OutputType {
-                    name: field.name().into(),
-                    ty: field_ty.kind(),
-                    src: db.source_code(field.loc().file_id()),
-                    definition: (offset, len).into(),
-                }))
-            }
-        } else if let Some(loc) = field.ty().loc() {
-            let field_ty_offset = loc.offset();
-            let field_ty_len = loc.node_len();
-            diagnostics.push(ApolloDiagnostic::UndefinedDefinition(UndefinedDefinition {
-                ty: field.ty().name(),
-                src: db.source_code(loc.file_id()),
-                definition: (field_ty_offset, field_ty_len).into(),
-            }))
-        } else {
-            diagnostics.push(ApolloDiagnostic::UndefinedDefinition(UndefinedDefinition {
-                ty: field.ty().name(),
-                src: db.source_code(field.loc().file_id()),
-                definition: (offset, len).into(),
-            }))
-        }
-    }
+    diagnostics.extend(db.validate_field_definitions(interface_def.fields_definition().to_vec()));
 
     let interfaces = db.interfaces();
     let defined_interfaces: HashSet<ValidationSet> = interfaces
