+++ crates/apollo-compiler/test_data/diagnostics/0033_object_type_definition_with_duplicate_fields.txt
@@ -20,7 +20,7 @@
                 ),
             },
             help: Some(
-                "`name` field must only be defined once in this object type definition.",
+                "`name` field must only be defined once in this definition.",
             ),
         },
     ),
